+++ Changelog.md
@@ -1,3 +1,21 @@
+# v1.3.3-dev
+
+Fixed:
+
+- Folder creation dialog wasn't working for some reason
+
+# v1.3.2-dev
+
+Fixed:
+
+- Text overflow on list item
+
+# v1.3.1-dev
+
+Fixed:
+
+- Back button on Testing screen
+
 # v1.3.0
 
 Added Features: